+++ web/jobs/assets/SettingsView-DoqgkZ3R.js
@@ -1,4 +1,4 @@
-import{c as L,r as a,j as e,P as H,a as ge,J as W,d as Be,N as J,D as G,O as M,m as F,u as De,C as le,M as ye,U as Ue,Q as oe,s as D,V as ke,W as Ge,X as K,Y as we,Z as Ee,$ as Ve,h as Fe,t as Ae,A as Ye,a0 as He,a1 as We,a2 as ve,a3 as Je,a4 as Ke,y as Qe,i as fe}from"./index-h3ha1Q1W.js";import{M as ce,C as Xe}from"./mail-BQr4u-9e.js";import{C as Te}from"./chevron-right-DVPxSih3.js";import{M as Ze}from"./monitor-up-B6ne_Rss.js";import{P as es}from"./pencil-C8YRIqx5.js";/**
+import{c as L,r as a,j as e,P as H,a as ge,J as W,d as Be,N as J,D as G,O as M,m as F,u as De,C as le,M as ye,U as Ue,Q as oe,s as D,V as ke,W as Ge,X as K,Y as we,Z as Ee,$ as Ve,h as Fe,t as Ae,A as Ye,a0 as He,a1 as We,a2 as ve,a3 as Je,a4 as Ke,y as Qe,i as fe}from"./index-BDp248s4.js";import{M as ce,C as Xe}from"./mail-DIFwT7Fw.js";import{C as Te}from"./chevron-right-rWuvWaHx.js";import{M as Ze}from"./monitor-up-CcQ7sqtC.js";import{P as es}from"./pencil-CICls6mj.js";/**
  * @license lucide-react v0.500.0 - ISC
  *
  * This source code is licensed under the ISC license.